+++ R/compare-opts.R
@@ -4,8 +4,9 @@ compare_opts <- function(...,
                          ignore_srcref = TRUE,
                          ignore_attr = FALSE,
                          ignore_encoding = TRUE,
+                         ignore_function_env = FALSE,
                          ignore_formula_env = FALSE,
-                         ignore_function_env = FALSE
+                         list_as_map = FALSE
                          ) {
 
   base <- old_opts(...)
@@ -16,8 +17,9 @@ compare_opts <- function(...,
     ignore_srcref = ignore_srcref,
     ignore_attr = ignore_attr,
     ignore_encoding = ignore_encoding,
+    ignore_function_env = ignore_function_env,
     ignore_formula_env = ignore_formula_env,
-    ignore_function_env = ignore_function_env
+    list_as_map = list_as_map
   )
 
   utils::modifyList(waldo, base)