+++ NEWS.md
@@ -4,6 +4,15 @@
   how the proxy has been used to change the object. This makes it easier to 
   see when and how a proxy is involved (#73).
 
+* You can use a `waldo_opts` attribute to now control how objects are compared.
+  This should be a list with the same names and valid values as the arguments
+  to `compare()` (@dmurdoch).
+
+* `compare()` gains a `list_as_map` argument thanks to an idea from @dmurdoch.
+  It allows you to compare the behaviour of two lists when they are used to 
+  connect names to values (i.e. the list is operating as a map or dictionary). 
+  It removes NULLs and sorts named components (#72).
+
 * `compare()` gains a basic row-by-row diff for data frames (#78) that is shown 
   when the values within a data frame are different, and for numeric matrices
   (#76).